admin: Enter debug mode
We can debug Django with django-debug-toolbar and django-debug-panel.
Besides the changes of this commit, we also need to do the following:

In sites-available/synnefo-ssl, add alias and proxy for `static2`. Also,
copy the files of django-debug-toolbar from /usr/lib/python2.7/... to
the `static2` folder.

NOTE: For the changes of this commit to take effect, we need to change
the INTERNAL_IPS value.