admin: Add check in `is_resource_useful` function
In order to check if a resource is useful, we must not only take into
account its limit, but also its current usage, if any. Therefore, we
extend the `is_resource_useful` function in order to accept a `usage`
parameter.

The usage parameter is useful in a scenario where the resource limit is
0 (e.g. due to a project termination) and someone still uses this
resource.