cyclades: Facilitate volume assignments
Add helper function `assign_volume_to_server` that facilitates the
assignment of a volume to a server. In a nutshell, this function works
in the DB level and ensures that the proper index will be assigned to
the volume and that the volume will be associated with the server.

Also, add the necessary tests for this new function.