cyclades: Implement creation of standalone volumes
Allow the creation of volumes that are not attached to any server. Such
volumes can be created only from detachable volume types.

The API changes for this feature are trivial; we simply do not require a
server id when creating a volume, which is after all compatible with the
OpenStack API. If the user does provide a server id after all, then we
retain backwards compatibility by creating and attaching this volume to
the server in one action.