cyclades: Implement attach volume action
The attach functionality is available for detachable volumes which are
in "AVAILABLE" status. In order to make the attach work, we send to
Ganeti an ExtStorage parameter called "reuse_data", which hints the
storage backend to not create a new disk but reuse an existing one,
provided that the volume has been initialized.

The API for the attach action adheres to the server_attachments
OpenStack API and **TODO**.