cyclades: Implement detach volume action
The detach functionality is available for detachable volumes which are
in "IN USE" status. In order to make the detach work, we send to Ganeti
a remove action and with that, a parameter called "keep_disks". This
parameter exists only in snf-ganeti and when passed to Ganeti, it skips
the removal of the disk.

The API for the detach action adheres to the server_attachments
OpenStack API and **TODO**.

Moreover, re-implement the delete functionality, which previously was in
the place of the detach action (since the volume was deleted after being
detached). The re-implemented delete function will have the same
capabilities as before, meaning that it will not be able to delete
detached volumes. This will be available in the following commits.