cyclades: Extend delete volume action
Extend the delete volume action to handle two more cases:

1) Uninitialized standalone volumes: These are volumes that have been
created in the DB but have not been created in the backend yet. This is
an easy case to handle, as we can delete the volume instantly.
2) Detached volumes: These are volumes that have been created in the
backend and are not currently attached to any server. In order to delete
these volumes, we must first attach them to a helper server and then
issue a delete.