cyclades: Add legacy_backend_volume_uuid field
The rationale for this change is explained in the design document of
detachable volumes, in the **TODO** section.

In a gist, volumes that were created prior to Archipelago version
**TODO** used the Ganeti parameter VOL_NAME as their name, instead of
VOL_CNAME. Since we need a common name for a volume across Ganeti
clusters and VOL_NAME changes after a detachment, we need a way to make
VOL_CNAME evaluate to the initial VOL_NAME of the volume. To do so, we
need an extra field ("legacy_backend_volume_uuid") to store that info.

Moreover, we also need a schema migration to create a new column in the
Cyclades DB, as well as a migration script that will get the necessary
info from Ganeti and store them in this field.